fix(DEVING-48): searchFilter 관련 로직 별도로 관리
